+++ spark/components/footer/react/SprkFooter.js
@@ -3,9 +3,9 @@ import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import { uniqueId } from 'lodash';
 import SprkIcon from '../../icons/react/SprkIcon';
-import SprkFooterGlobalSection from './SprkFooterGlobalSection';
-import SprkFooterConnectIcons from './SprkFooterConnectIcons';
-import SprkFooterAwards from './SprkFooterAwards';
+import SprkFooterGlobalSection from './components/SprkFooterGlobalSection/SprkFooterGlobalSection';
+import SprkFooterConnectIcons from './components/SprkFooterConnectIcons/SprkFooterConnectIcons';
+import SprkFooterAwards from './components/SprkFooterAwards/SprkFooterAwards';
 
 class SprkFooter extends Component {
   constructor(props) {
