+++ spark/components/footer/react/components/SprkFooterAwards/SprkFooterAwards.test.js
