+++ spark/components/footer/react/components/SprkFooterGlobalSection/SprkFooterGlobalSection.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { uniqueId } from 'lodash';
 import PropTypes from 'prop-types';
-import SprkIcon from '../../icons/react/SprkIcon';
+import SprkIcon from '../../../../icons/react/SprkIcon';
 
 class SprkFooterGlobalSection extends Component {
   constructor(props) {
